Refactoring; split `tests/forms/test_permission` test suite into `tests/forms/test_permission_register` and `tests/forms/test_permission_edit.py`
